Added built-in functions. Modified clipboard control
Built-in Functions:
[DATE], [TIME], [TIMESTAMP]

Clipboard:
Added a delay (0.5 seconds) before setting the clipboard content back
to whatever was in there before the script ran. It doesn't seem to work
with objects (files, etc.), wehn setting the clipboard back to its
original content it seems to be coerced into a string, instead of
whatever the original object type was.